+++ relay-server/src/services/buffer/stack_provider/sqlite.rs
@@ -23,8 +23,8 @@ pub struct SqliteStackProvider {
 #[warn(dead_code)]
 impl SqliteStackProvider {
     /// Creates a new [`SqliteStackProvider`] from the provided [`Config`].
-    pub async fn new(config: &Config) -> Result<Self, SqliteEnvelopeStoreError> {
-        let envelope_store = SqliteEnvelopeStore::prepare(config).await?;
+    pub async fn new(partition_id: u8, config: &Config) -> Result<Self, SqliteEnvelopeStoreError> {
+        let envelope_store = SqliteEnvelopeStore::prepare(partition_id, config).await?;
         Ok(Self {
             envelope_store,
             batch_size_bytes: config.spool_envelopes_batch_size_bytes(),
@@ -145,7 +145,7 @@ mod tests {
     #[tokio::test]
     async fn test_flush() {
         let config = mock_config();
-        let mut stack_provider = SqliteStackProvider::new(&config).await.unwrap();
+        let mut stack_provider = SqliteStackProvider::new(0, &config).await.unwrap();
 
         let own_key = ProjectKey::parse("a94ae32be2584e0bbd7a4cbb95971fee").unwrap();
         let sampling_key = ProjectKey::parse("b81ae32be2584e0bbd7a4cbb95971fe1").unwrap();